+++ fuzz/bit_manip/bit_manip_fuzzer_32bit_input.c
@@ -177,11 +177,11 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
     if (has_single_bit_ui_result != (val1 != 0U && (val1 & (val1 - 1U)) == 0U)) __builtin_trap();
 
     // Fuzzing get_req_bit_width_ui
-    uint8_t get_req_bit_width_ui_result = get_req_bit_width_ui(val1);
+    uint32_t get_req_bit_width_ui_result = get_req_bit_width_ui(val1);
     if (get_req_bit_width_ui_result != (UINT_WIDTH - (val1 == 0U ? UINT_WIDTH : count_leading_zeros_ui(val1)))) __builtin_trap();
 
     // Fuzzing bit_floor_ui
-    uint8_t bit_floor_ui_result = bit_floor_ui(val1);
+    uint32_t bit_floor_ui_result = bit_floor_ui(val1);
     if (bit_floor_ui_result != (val1 == 0U ? 0U : 1U << (UINT_WIDTH - 1U - count_leading_zeros_ui(val1)))) __builtin_trap();
 
     return 0;
